code/flask/requirements.txt
Flask
youtube-transcript-api
openai
python-dotenv